refactor mapper_builder
