+++ client/starwhale/base/artifact/__init__.py
@@ -1,0 +1,4 @@
+from .base import AsyncArtifactWriterBase
+from .rotated_bin import RotatedBinWriter
+
+__all__ = ["RotatedBinWriter", "AsyncArtifactWriterBase"]
